Fixed many compiler warnings from g++ 4.4

+++ src/pyinterp.cc
@@ -266,7 +266,7 @@ value_t python_interpreter_t::python_command(call_scope_t& args)
     std::strcpy(argv[i + 1], arg.c_str());
   }
 
-  int status = Py_Main(args.size() + 1, argv);
+  int status = Py_Main(static_cast<int>(args.size()) + 1, argv);
 
   for (std::size_t i = 0; i < args.size() + 1; i++)
     delete[] argv[i];
